+++ src/libponyrt/ds/fun.c
@@ -65,14 +65,14 @@ static uint64_t siphash24(const unsigned char* key, const char* in, size_t len)
   return v0 ^ v1 ^ v2 ^ v3;
 }
 
-size_t hash_block(const void* p, size_t len)
+uint64_t hash_block(const void* p, size_t len)
 {
-  return (size_t)siphash24(the_key, (const char*)p, len);
+  return siphash24(the_key, (const char*)p, len);
 }
 
-size_t hash_str(const char* str)
+uint64_t hash_str(const char* str)
 {
-  return (size_t)siphash24(the_key, str, strlen(str));
+  return siphash24(the_key, str, strlen(str));
 }
 
 size_t hash_ptr(const void* p)
